bsns-mcp/requirements.txt

flask
newspaper3k
sumy